+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project follows [Semantic Versioning](https://semver.org/).
 
 ## [1.3.0]
-- Add typing
+- Add typing: https://peps.python.org/pep-0561/
+- Changed minimum Python version to 3.9
 
 ## [1.2.2]
 - On a non-secure connection still generate a login event on got_connection